Started implementing a renderer for drawing things in world space since using imgui is too slow and limiting

+++ ProjectValkyrie/ValkyrieDLL/GameHud.cpp
@@ -22,5 +22,5 @@ void GameHud::ReadFromBaseAddress(int baseAddr)
 
 bool GameHud::WasChatOpenMillisAgo(int millis)
 {
-	return (GetTickCount() - chatLastOpenTimestamp) < millis;
+	return ((int)GetTickCount() - chatLastOpenTimestamp) < millis;
 }
